+++ slack-sup/api/presenters/sup_presenter.rb
@@ -8,6 +8,7 @@ module Api
 
       property :id, type: String, desc: "S'Up ID."
       property :outcome, type: String, desc: "S'up outcome."
+      property :captain_user_name, type: String, desc: 'Captain user name.'
       property :created_at, type: DateTime, desc: "Date/time when the S'Up was created."
       property :updated_at, type: DateTime, desc: "Date/time when the S'Up was updated."